requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.0
gradio_client==1.8.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langsmith==0.3.31
markdownify==0.14.1
monotonic==1.6
MouseInfo==0.1.3
numpy==2.2.4
ollama==0.4.8
openai==1.68.2
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pillow==11.1.0
playwright==1.51.0
posthog==3.25.0
psutil==7.0.0
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PyGetWindow==0.0.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
zstandard==0.23.0
